[hw08-websocket] Added WsService to backend
