+++ node-2023-11/hw04_les07-webServers/package.json
@@ -19,6 +19,7 @@
     "passport": "^0.7.0",
     "passport-http": "^0.3.0",
     "pug": "2.0.0-beta11",
+    "ws": "^8.16.0",
     "yup": "^1.3.3"
   },
   "devDependencies": {
